requirements.txt
streamlit>=1.28.0
pandas>=2.0.0
sqlalchemy>=2.0.0
pyodbc>=4.0.0
faiss-cpu>=1.7.4
google-generativeai>=0.3.0